Feature: Enable AirPlay route picker where available
This adds an additional button to the player toolbar which brings up the system dialog to select the AirPlay destination.

This also fixes a long stand bug that only disabled the chapter button instead of actually removing it (which came to be because I didn't understand `isEnabled` at the time).

This also introduces an animated change in the toolbar when the sleep timer is started or ended.

Signed-off-by: Florian Pichler <fp@ylk.gd>